backend/requirements.txt
fastapi==0.79.0
motor==3.0.0
python-dotenv==0.20.0
python-decouple==3.6
pydantic[email]
